+++ test/parallel/test-repl-autolibs.js
@@ -21,11 +21,12 @@
 
 'use strict';
 const common = require('../common');
+const ArrayStream = require('../common/arraystream');
 const assert = require('assert');
 const util = require('util');
 const repl = require('repl');
 
-const putIn = new common.ArrayStream();
+const putIn = new ArrayStream();
 repl.start('', putIn, null, true);
 
 test1();
