+++ test/parallel/test-repl-domain.js
@@ -20,11 +20,12 @@
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 'use strict';
-const common = require('../common');
+require('../common');
+const ArrayStream = require('../common/arraystream');
 
 const repl = require('repl');
 
-const putIn = new common.ArrayStream();
+const putIn = new ArrayStream();
 repl.start('', putIn);
 
 putIn.write = function(data) {
