+++ test/parallel/test-repl-eval-scope.js
@@ -1,10 +1,11 @@
 'use strict';
 const common = require('../common');
+const ArrayStream = require('../common/arraystream');
 const assert = require('assert');
 const repl = require('repl');
 
 {
-  const stream = new common.ArrayStream();
+  const stream = new ArrayStream();
   const options = {
     eval: common.mustCall((cmd, context) => {
       assert.strictEqual(cmd, '.scope\n');
