+++ test/parallel/test-repl-options.js
@@ -21,11 +21,12 @@
 
 'use strict';
 const common = require('../common');
+const ArrayStream = require('../common/arraystream');
 const assert = require('assert');
 const repl = require('repl');
 
 // Create a dummy stream that does nothing
-const stream = new common.ArrayStream();
+const stream = new ArrayStream();
 
 // 1, mostly defaults
 const r1 = repl.start({
