+++ test/parallel/test-repl-pretty-stack.js
@@ -1,5 +1,6 @@
 'use strict';
-const common = require('../common');
+require('../common');
+const ArrayStream = require('../common/arraystream');
 const fixtures = require('../common/fixtures');
 const assert = require('assert');
 const repl = require('repl');
@@ -8,8 +9,8 @@ const repl = require('repl');
 function run({ command, expected }) {
   let accum = '';
 
-  const inputStream = new common.ArrayStream();
-  const outputStream = new common.ArrayStream();
+  const inputStream = new ArrayStream();
+  const outputStream = new ArrayStream();
 
   outputStream.write = (data) => accum += data.replace('\r', '');
