+++ test/parallel/test-repl-reset-event.js
@@ -21,7 +21,7 @@
 
 'use strict';
 const common = require('../common');
-
+const ArrayStream = require('../common/arraystream');
 const assert = require('assert');
 const repl = require('repl');
 const util = require('util');
@@ -29,7 +29,7 @@ const util = require('util');
 common.allowGlobals(42);
 
 // Create a dummy stream that does nothing
-const dummy = new common.ArrayStream();
+const dummy = new ArrayStream();
 
 function testReset(cb) {
   const r = repl.start({
