+++ test/parallel/test-repl-syntax-error-stack.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const common = require('../common');
+const ArrayStream = require('../common/arraystream');
 const fixtures = require('../common/fixtures');
 const assert = require('assert');
 const repl = require('repl');
@@ -10,7 +11,7 @@ process.on('exit', () => {
   assert.strictEqual(found, true);
 });
 
-common.ArrayStream.prototype.write = function(output) {
+ArrayStream.prototype.write = function(output) {
   // Matching only on a minimal piece of the stack because the string will vary
   // greatly depending on the JavaScript engine. V8 includes `;` because it
   // displays the line of code (`var foo bar;`) that is causing a problem.
@@ -20,7 +21,7 @@ common.ArrayStream.prototype.write = function(output) {
     found = true;
 };
 
-const putIn = new common.ArrayStream();
+const putIn = new ArrayStream();
 repl.start('', putIn);
 let file = fixtures.path('syntax', 'bad_syntax');
