+++ test/parallel/test-repl-tab-complete.js
@@ -22,6 +22,7 @@
 'use strict';
 
 const common = require('../common');
+const ArrayStream = require('../common/arraystream');
 const assert = require('assert');
 const fixtures = require('../common/fixtures');
 const hasInspector = process.config.variables.v8_enable_inspector === 1;
@@ -44,7 +45,7 @@ function getNoResultsFunction() {
 }
 
 const works = [['inner.one'], 'inner.o'];
-const putIn = new common.ArrayStream();
+const putIn = new ArrayStream();
 const testMe = repl.start('', putIn);
 
 // Some errors are passed to the domain, but do not callback
@@ -536,7 +537,7 @@ testCustomCompleterAsyncMode.complete('a', common.mustCall((error, data) => {
 }));
 
 // tab completion in editor mode
-const editorStream = new common.ArrayStream();
+const editorStream = new ArrayStream();
 const editor = repl.start({
   stream: editorStream,
   terminal: true,
@@ -559,7 +560,7 @@ editor.completer('var log = console.l', common.mustCall((error, data) => {
 
 {
   // tab completion of lexically scoped variables
-  const stream = new common.ArrayStream();
+  const stream = new ArrayStream();
   const testRepl = repl.start({ stream });
 
   stream.run([`
